requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.11.4

chromadb==0.6.3
openai==1.40.0
anthropic==0.34.0

neo4j==5.23.1

langchain==0.3.0
langchain-openai==0.2.0
langchain-google-genai==2.0.0
langchain-upstage==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langgraph==0.2.28

fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

ragas==0.2.6
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2